+++ engine/spec/components/citizens_advice_components/table_spec.rb
@@ -1,55 +1,79 @@
 # frozen_string_literal: true
 
 RSpec.describe CitizensAdviceComponents::Table, type: :component do
-  subject(:component) do
-    render_inline(
+  let(:sample_header) do
+    ["Your location", "Post box collection times"]
+  end
+
+  let(:sample_rows) do
+    [
+      ["City or town", "9am to 6.30pm"],
+      ["Areas with lots of businesses - known as commercial", "9am to 7.30pm"],
+      ["Very rural areas - for example, where there aren't many people", "9am to 4pm"],
+      ["Rest of the UK", "9am to 5.30pm"]
+    ]
+  end
+
+  context "when missing headers" do
+    subject(:component) do
+      described_class.new(header: [], rows: sample_rows)
+    end
+
+    it "does not render" do
+      render_inline component
+      expect(rendered_component).to have_no_table
+    end
+  end
+
+  context "when missing rows" do
+    subject(:component) do
+      described_class.new(header: sample_header, rows: [])
+    end
+
+    it "does not render" do
+      render_inline component
+      expect(rendered_component).to have_no_table
+    end
+  end
+
+  context "when valid table" do
+    subject(:component) do
       described_class.new(
-        header: header,
-        rows: rows,
-        caption: caption.presence
+        header: sample_header,
+        rows: sample_rows
       )
-    )
+    end
+
+    it "matches snapshot" do
+      expect(render_inline(component).to_html).to match_snapshot("table")
+    end
   end
 
-  let(:header) do
-    [
-      "Your location",
-      "Post box collection times"
-    ]
+  context "when a caption is present" do
+    subject(:component) do
+      described_class.new(
+        header: sample_header,
+        rows: sample_rows,
+        caption: "Example caption"
+      )
+    end
+
+    it "renders a caption" do
+      render_inline component
+      expect(rendered_component).to have_css "caption", text: "Example caption"
+    end
   end
 
-  let(:rows) do
-    [
-      [
-        "City or town",
-        "9am to 6.30pm"
-      ],
-      [
-        "Areas with lots of businesses - known as commercial",
-        "9am to 7.30pm"
-      ],
-      [
-        "Very rural areas - for example, where there aren't many people",
-        "9am to 4pm"
-      ],
-      [
-        "Rest of the UK",
-        "9am to 5.30pm"
-      ]
-    ]
-  end
+  context "when additional empty rows" do
+    subject(:component) do
+      described_class.new(
+        header: sample_header,
+        rows: sample_rows.concat([["", ""], ["", ""], ["", ""]])
+      )
+    end
 
-  let(:caption) { "I am a nice table" }
-
-  it "renders correctly" do
-    expect(component.to_html).to match_snapshot("table")
-  end
-
-  context "when there is no caption" do
-    let(:caption) { nil }
-
-    it "does not render a caption" do
-      expect(component.css("table caption")).to be_empty
+    it "strips empty rows" do
+      expect(render_inline(component).to_html).to match_snapshot("table")
     end
   end
 end