+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - 🥾 Footer: Supports dynamic copyright year
 
+**Bugfixes**
+
+- 🧂 Tables: Added render guards for missing headers and empty rows
 
 ## <sub>v4.3.0-alpha.0</sub>